requirements.txt
cassandra-driver==3.25.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
Flask-Cors==3.0.10
geomet==0.2.1.post1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pipreqs==0.4.11
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
yarg==0.1.9
